Branch 154257269

+++ tensorflow/python/debug/wrappers/framework_test.py
@@ -17,8 +17,10 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+import os
 import shutil
 import tempfile
+import threading
 
 import numpy as np
 
@@ -39,7 +41,7 @@ from tensorflow.python.platform import googletest
 class TestDebugWrapperSession(framework.BaseDebugWrapperSession):
   """A concrete implementation of BaseDebugWrapperSession for test."""
 
-  def __init__(self, sess, dump_root, observer):
+  def __init__(self, sess, dump_root, observer, thread_name_filter=None):
     # Supply dump root.
     self._dump_root = dump_root
 
@@ -47,7 +49,8 @@ class TestDebugWrapperSession(framework.BaseDebugWrapperSession):
     self._obs = observer
 
     # Invoke superclass constructor.
-    framework.BaseDebugWrapperSession.__init__(self, sess)
+    framework.BaseDebugWrapperSession.__init__(
+        self, sess, thread_name_filter=thread_name_filter)
 
   def on_session_init(self, request):
     """Override abstract on-session-init callback method."""
@@ -155,7 +158,7 @@ class DebugWrapperSessionTest(test_util.TensorFlowTestCase):
     self._c_val = np.array([[-4.0], [6.0]])
 
     self._a_init = constant_op.constant(
-        self._a_init_val, shape=[2, 2], name="a1_init")
+        self._a_init_val, shape=[2, 2], name="a_init")
     self._b_init = constant_op.constant(
         self._b_init_val, shape=[2, 1], name="b_init")
 
@@ -180,7 +183,8 @@ class DebugWrapperSessionTest(test_util.TensorFlowTestCase):
 
   def tearDown(self):
     # Tear down temporary dump directory.
-    shutil.rmtree(self._dump_root)
+    if os.path.isdir(self._dump_root):
+      shutil.rmtree(self._dump_root)
 
     ops.reset_default_graph()
 
@@ -320,6 +324,65 @@ class DebugWrapperSessionTest(test_util.TensorFlowTestCase):
     wrapper = TestDebugWrapperSession(self._sess, self._dump_root,
                                       self._observer)
     wrapper.close()
+
+  def testWrapperThreadNameFilterMainThread(self):
+    wrapper = TestDebugWrapperSession(
+        self._sess, self._dump_root, self._observer,
+        thread_name_filter="MainThread")
+
+    child_run_output = []
+    def child_thread_job():
+      child_run_output.append(wrapper.run(self._b_init))
+
+    thread = threading.Thread(name="ChildThread", target=child_thread_job)
+    thread.start()
+    self.assertAllClose(self._a_init_val, wrapper.run(self._a_init))
+    thread.join()
+    self.assertAllClose([self._b_init_val], child_run_output)
+
+    dump = debug_data.DebugDumpDir(self._dump_root)
+    self.assertEqual(1, dump.size)
+    self.assertEqual("a_init", dump.dumped_tensor_data[0].node_name)
+
+  def testWrapperThreadNameFilterChildThread(self):
+    wrapper = TestDebugWrapperSession(
+        self._sess, self._dump_root, self._observer,
+        thread_name_filter=r"Child.*")
+
+    child_run_output = []
+    def child_thread_job():
+      child_run_output.append(wrapper.run(self._b_init))
+
+    thread = threading.Thread(name="ChildThread", target=child_thread_job)
+    thread.start()
+    self.assertAllClose(self._a_init_val, wrapper.run(self._a_init))
+    thread.join()
+    self.assertAllClose([self._b_init_val], child_run_output)
+
+    dump = debug_data.DebugDumpDir(self._dump_root)
+    self.assertEqual(1, dump.size)
+    self.assertEqual("b_init", dump.dumped_tensor_data[0].node_name)
+
+  def testWrapperThreadNameFilterBothThreads(self):
+    wrapper = TestDebugWrapperSession(
+        self._sess, self._dump_root, self._observer,
+        thread_name_filter=None)
+
+    child_run_output = []
+    def child_thread_job():
+      child_run_output.append(wrapper.run(self._b_init))
+
+    thread = threading.Thread(name="ChildThread", target=child_thread_job)
+    thread.start()
+    self.assertAllClose(self._a_init_val, wrapper.run(self._a_init))
+    thread.join()
+    self.assertAllClose([self._b_init_val], child_run_output)
+
+    dump = debug_data.DebugDumpDir(self._dump_root, validate=False)
+    self.assertEqual(2, dump.size)
+    self.assertItemsEqual(
+        ["a_init", "b_init"],
+        [datum.node_name for datum in dump.dumped_tensor_data])
 
   def testUsingNonDirectSessionRaisesNotImplementedError(self):
     # TODO(cais): Remove this test once tfdbg is integrated with GrpcSession.
